Fix more tests
