+++ Cognite.Simulator.Extensions/SequencesExtensions.cs
@@ -415,105 +415,6 @@ namespace Cognite.Simulator.Extensions
             }
             return result.ToArray();
         }
-
-        // /// <summary>
-        // /// Simply upserts a key value pair into a key value pair based sequence
-        // /// </summary>
-        // /// <param name="sequences">CDF Sequence resource</param>
-        // /// <param name="sequenceExternalId">External id of the sequence.</param>
-        // /// <param name="updateKey">Key in sequence to update</param>
-        // /// <param name="updateValue">Value to update the Key value pair to</param>
-        // /// <param name="token">The cancellation token</param>
-        // /// <returns>Nothing</returns>
-        // public static async 
-        // Task UpsertItemInKVPSequence(this SequencesResource sequences, string sequenceExternalId, string updateKey, string updateValue, CancellationToken token) {
-        //     if (sequenceExternalId == "") {
-        //         throw new Exception("External Id is required to upsert a sequence");
-        //     }
-        //     SequenceRowQuery query = new SequenceRowQuery();
-        //     query.ExternalId = sequenceExternalId;
-        //     var sequenceList = await sequences.ListRowsAsync(query, token).ConfigureAwait(false);;
-        //     var rows = sequenceList.Rows;
-        //     var columns = sequenceList.Columns;
-        //     var newRows = rows.Select(r => new SequenceRow
-        //         {
-        //             RowNumber = r.RowNumber,
-        //             Values = r.Values
-        //         }).ToList();
-        //     // Checking to see if the key alread exists
-        //     var foundIndex = newRows.FindIndex(0, newRows.Count, (item => item.Values.FirstOrDefault().ToString().Contains(updateKey) ));
-        //     if (foundIndex == -1){
-        //         newRows.Insert(newRows.Count , new SequenceRow{
-        //             RowNumber = newRows.Count,
-        //             Values = new List<MultiValue> {
-        //                 MultiValue.Create(updateKey),
-        //                 MultiValue.Create(updateValue)
-        //             }
-        //         });
-        //     } else {
-        //         newRows[foundIndex].Values = new List<MultiValue> {
-        //             MultiValue.Create(updateKey),
-        //             MultiValue.Create(updateValue)
-        //         };
-        //     }
-
-        //     SequenceDataCreate sequenceData = new SequenceDataCreate
-        //     {
-        //         ExternalId = sequenceExternalId,
-        //         Columns = new List<string> {
-        //                     KeyValuePairSequenceColumns.Key,
-        //                     KeyValuePairSequenceColumns.Value
-        //                 },
-        //         Rows = newRows
-        //     };
-        //     var rowsToCreate = new List<SequenceDataCreate>();
-        //     rowsToCreate.Add(sequenceData);
-        //     var result = await sequences.InsertAsync(
-        //         rowsToCreate,
-        //         keyChunkSize: 10,
-        //         valueChunkSize: 100,
-        //         sequencesChunk: 10,
-        //         throttleSize: 1,
-        //         RetryMode.None,
-        //         SanitationMode.None,
-        //         token
-        //     ).ConfigureAwait(false);
-                
-        //     if (!result.IsAllGood)
-        //     {
-        //         throw new SimulatorIntegrationSequenceException("Could not update simulator integration sequence", result.Errors);
-        //     }
-        // }
-
-    //     /// <summary>
-    //     /// Gets the externalId of a SimulatorIntegrations sequence (This function can throw an exception)
-    //     /// </summary>
-    //     /// <param name="sequences">CDF Sequence resource</param>
-    //     /// <param name="simulatorName">The simulator name.</param>
-    //     /// <param name="simulatorDataSetId">The simulator datasetId</param>
-    //     /// <param name="connectorName">The connector Name</param>
-    //     /// <param name="token">The cancellation token</param>
-    //     /// <returns>externalId string</returns>
-    //     public static async Task<string> GetSequenceExternalId(this SequencesResource sequences, string simulatorName, long simulatorDataSetId, string connectorName, CancellationToken token) {
-    //         var simulatorSequenceIds = new Dictionary<string, string>();
-    //         var simulatorsDict = new SimulatorIntegration
-    //             {
-    //                 Simulator = simulatorName,
-    //                 DataSetId = simulatorDataSetId,
-    //                 ConnectorName = connectorName
-    //             };
-            
-    //         var integrations = await sequences.GetOrCreateSimulatorIntegrations(
-    //                 new List<SimulatorIntegration> { simulatorsDict },
-    //                 token).ConfigureAwait(false);
-    //         foreach (var integration in integrations)
-    //         {
-    //             simulatorSequenceIds.Add(
-    //                 integration.Metadata[BaseMetadata.SimulatorKey],
-    //                 integration.ExternalId);
-    //         }
-    //         return simulatorSequenceIds[simulatorName];                
-    //     }
     }
 
     /// <summary>
