+++ Cognite.Simulator.Extensions/Types.cs
@@ -374,58 +374,4 @@ namespace Cognite.Simulator.Extensions
             ExternalIdOverwrite = externalId;
         }
     }
-
-    // /// <summary>
-    // /// Represents simulator integration information in CDF. Each connector can register one or more
-    // /// simulator integration record with CDF
-    // /// </summary>
-    // public class SimulatorIntegration
-    // {
-    //     /// <summary>
-    //     /// Name of the simulator
-    //     /// </summary>
-    //     public string Simulator { get; set; }
-        
-    //     /// <summary>
-    //     /// Name of the connector
-    //     /// </summary>
-    //     public string ConnectorName { get; set; }
-        
-    //     /// <summary>
-    //     /// ID of the data set that holds simulator data in CDF
-    //     /// </summary>
-    //     public long? DataSetId { get; set; }
-    // }
-
-    // /// <summary>
-    // /// Represents data used to update the simulator integration information in CDF
-    // /// </summary>
-    // public class SimulatorIntegrationUpdate : SimulatorIntegration
-    // {
-    //     /// <summary>
-    //     /// Version of the connector
-    //     /// </summary>
-    //     public string ConnectorVersion { get; set; }
-        
-    //     /// <summary>
-    //     /// Version of the simulator
-    //     /// </summary>
-    //     public string SimulatorVersion { get; set; }
-        
-    //     /// <summary>
-    //     /// Flag indicating if the connector is using Cognite's Simulator Integration API
-    //     /// or just core CDF resources.
-    //     /// </summary>
-    //     public bool SimulatorApiEnabled { get; set; }
-        
-    //     /// <summary>
-    //     /// Flag indicating if there connector will check for simulator liceses
-    //     /// </summary>
-    //     public bool LicenseCheckEnabled { get; set; }
-
-    //     /// <summary>
-    //     /// Any extra information that can be registered by the connector
-    //     /// </summary>
-    //     public Dictionary<string, string> ExtraInformation { get; set; } = new Dictionary<string, string>();
-    // }
 }
